created FriendContext and adjusted usage accordingly
