+++ src/index.js
@@ -3,11 +3,14 @@ import ReactDOM from 'react-dom/client'
 import './index.css'
 import Root from './App'
 import { UserProvider } from './context/UserContext'
+import { FriendProvider } from './context/FriendContext'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <UserProvider>
-      <Root />
+      <FriendProvider>
+        <Root />
+      </FriendProvider>
     </UserProvider>
   </React.StrictMode>
 )
